% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{calcLogRatio}
\alias{calcLogRatio}
\title{Calculate all pairwise log-ratios}
\usage{
calcLogRatio(df)
}
\arguments{
\item{df}{a data.frame where the first column contains class labels and the remaining p columns are features}
}
\value{
A data frame with all choose(p,2) log ratios
}
\description{
Compute all pairwise log-ratios froma  data.frame where rows=samples and columns=features/parts
}
\examples{
calcLogRatio()

}
\references{

}
