% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateMetrics.R
\name{etest2}
\alias{etest2}
\title{Wrapper around the Energy package Multisample E-statistic (Energy) Test of Equal Distributions}
\usage{
etest2(
  x,
  sizes,
  distance = FALSE,
  method = c("original", "discoB", "discoF"),
  R
)
}
\arguments{
\item{x}{sorted matrix with n-samples}

\item{sizes}{class sizes}

\item{distance}{is a distance matrix}

\item{method}{method}

\item{R}{number of permuattions}
}
\value{
A data frame with all choose(p,2) log ratios
}
\description{
Modifies the energy::eqdist.etest to return permuation distribution
}
\examples{
etest2()

}
\references{
Szekely, G. J. and Rizzo, M. L. (2004) Testing for Equal Distributions in High Dimension, InterStat, November (5).
M. L. Rizzo and G. J. Szekely (2010). DISCO Analysis: A Nonparametric Extension of Analysis of Variance, Annals of Applied Statistics, Vol. 4, No. 2, 1034-1055. http://dx.doi.org/10.1214/09-AOAS245
Szekely, G. J. (2000) Technical Report 03-05: E-statistics: Energy of Statistical Samples, Department of Mathematics and Statistics, Bowling Green State University.
}
\seealso{
[energy::eqdist.etest()]
}
