% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{fastImputeZeroes}
\alias{fastImputeZeroes}
\title{Impute zeroes}
\usage{
fastImputeZeroes(df.cdata2, impFactor = 1e-11)
}
\arguments{
\item{df.cdata2}{count or relative abundance matrix where rows are samples and columns are features}

\item{impFactor}{delta imputed value}
}
\value{
A matrix with multiplicatively imputed zeroes
}
\description{
Impute zeroes using a lightly modified multiplicative replacement strategy (Martin-Fernandez (2003))
}
\examples{
fastImputeZeroes()

}
\references{
Martin-Fernandez, J. A., Barcelo-Vidal, C., & Pawlowsky-Glahn, V. (2003). Dealing with Zeros and Missing Values in Compositional Data Sets Using Nonparametric Imputation. Mathematical Geology, 26.
}
