% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateMetrics.R
\name{featureSlectionPerformance}
\alias{featureSlectionPerformance}
\title{Mulitvariate Statistics of a set of log-ratios}
\usage{
featureSlectionPerformance(
  tbl,
  plot_ = F,
  nreps_energy = 1e+05,
  Method_Name = "Feat_Selection",
  scaleData = T,
  scenario = "Scenario_X",
  combinedF_reps = 5000
)
}
\arguments{
\item{tbl}{sorted matrix with n-samples}

\item{plot_}{class sizes}

\item{nreps_energy}{is a distance matrix}

\item{Method_Name}{method}

\item{scaleData}{a data.frame where the first column contains class labels and the remaining p columns are features}

\item{scenario}{a data.frame where the first column contains class labels and the remaining p columns are features}

\item{combinedF_reps}{a data.frame where the first column contains class labels and the remaining p columns are features}
}
\value{
A data frame with all choose(p,2) log ratios
}
\description{
Computes various multivariate statistics to be used in permutation testing
}
\examples{
featureSlectionPerformance()

}
\references{
Szekely, G. J. and Rizzo, M. L. (2004) Testing for Equal Distributions in High Dimension, InterStat, November (5).
M. L. Rizzo and G. J. Szekely (2010). DISCO Analysis: A Nonparametric Extension of Analysis of Variance, Annals of Applied Statistics, Vol. 4, No. 2, 1034-1055. http://dx.doi.org/10.1214/09-AOAS245
Szekely, G. J. (2000) Technical Report 03-05: E-statistics: Energy of Statistical Samples, Department of Mathematics and Statistics, Bowling Green State University.
}
