% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{getFeats_bySparsity}
\alias{getFeats_bySparsity}
\title{Get non-sparse feature}
\usage{
getFeats_bySparsity(compmat, mxSparsePercent = 0.5, ncountRequired = 1)
}
\arguments{
\item{compmat}{a matrix/data.frame with counts (sample x features)}

\item{mxSparsePercent}{thresold for percent of samples a feature must be present with at least 1 count}
}
\value{
A vector of colnames of non sparse features to keep
}
\description{
Computes names of features present above a thresold
}
\examples{
getFeats_bySparsity()

}
\references{

}
