% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{getLogRatios}
\alias{getLogRatios}
\title{Calculate select log-ratios}
\usage{
getLogRatios(z, ratioList, weighted = F)
}
\arguments{
\item{z}{a matrix/data.frame with imputed relative abundace data (sample x features)}

\item{ratioList}{a data.frame with columns eqaul to the numerator of log-ratio, denominator of log-ratio, and the log-ratio names}
}
\value{
A data frame with simulated data from two classes
}
\description{
Calculate specific log-ratios defined in a matrix where rows are ratio and columns describe the numerator, denominator and log-ratio names
}
\examples{
getLogRatios()

}
\references{

}
