% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{processCompData}
\alias{processCompData}
\title{Wrapper function for removing spasre features from count data prior to log ratio analysis}
\usage{
processCompData(tbl, minPrevalence = 0.9, permuteLabel = F)
}
\arguments{
\item{tbl}{a n-samples x p-features data.frame with count or relative abundance data}

\item{minPrevalence}{class labels for data}

\item{permuteLabel}{should lass labels be permuted}
}
\value{
A list containing:\tabular{ll}{
   \code{processedData} \tab a data.frame where the first column contains class labels and the remaining p columns are features \cr
   \tab \cr
   \code{impFactor} \tab A vector of class lables \cr
   \tab \cr
   \code{minClss} \tab A vector of class lables \cr
   \tab \cr
   \code{majClass} \tab A vector of class lables \cr
}
}
\description{
Removes sparse features based on a provided threshold,computes \eqn{\delta} impute factor for zero impuation and returns the minority and majority class for binary group data
}
\examples{
processCompData()

}
\references{

}
