% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{sampleAddLogisticNormal}
\alias{sampleAddLogisticNormal}
\title{Sample from Additive Logisitcs Normal}
\usage{
sampleAddLogisticNormal(n, dims_, mu, sigma, sigmaScale = 1, sampleName)
}
\arguments{
\item{n}{number of samples}

\item{dims_}{number of features/dimensions}

\item{mu}{mean vector of length 'dims_'}

\item{sigma}{dims_ x dims_ covaraince matrix}

\item{sigmaScale}{scale factor for covaraince matrix}

\item{sampleName}{class names for data}
}
\value{
A data frame with simulated data
}
\description{
Simulate data from additive logistics normal distribution (see Aitichinson (1986))
}
\examples{
\dontrun{
sampleAddLogisticNormal(n = 40,dims_ = 75,mu = rep(0,75) ,
  sigma = diag(1,75),sigmaScale = 1,sampleName = "S1")
}

}
\references{
Aitchinson, J 1986
}
