% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{sampleDirichlet}
\alias{sampleDirichlet}
\title{Sample from Additive Logisitcs Normal}
\usage{
sampleDirichlet(n1, dims, a1, sampleName)
}
\arguments{
\item{n1}{number of samples}

\item{dims}{number of features/dimensions}

\item{a1}{alpha vector of dirichiliet distribution}

\item{sampleName}{class names for data}
}
\value{
A data frame with simulated data
}
\description{
Simulate data from additive logistics normal distribution (see Aitichinson (1986))
}
\examples{
sampleDirichlet()

}
\references{
Aitchinson, J 1986
}
