% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionEnergyPermutation.R
\name{selectionEnergy.scaled}
\alias{selectionEnergy.scaled}
\title{Selection Energy Permutation}
\usage{
selectionEnergy.scaled(
  inputData,
  optimizationMetric = NULL,
  dcv_useInfoGain = T,
  dcv_nfold = 5,
  dcv_numRepeats = 1,
  nreps_energy = 10000,
  eps = 1e-08,
  targetFeats = NULL,
  patience = 25,
  alpha_ = 0.05
)
}
\arguments{
\item{inputData}{xxx}

\item{optimizationMetric}{xxx}

\item{dcv_useInfoGain}{xxx}

\item{dcv_nfold}{xxx}

\item{dcv_numRepeats}{xxx}

\item{nreps_energy}{xxx}

\item{eps}{xxx}

\item{targetFeats}{xxx}

\item{patience}{xxx}

\item{alpha_}{xxx}
}
\value{
A list containing:\tabular{ll}{
   \code{lrs} \tab a p-row log-ratio scoring matrix aggregated across nfolds and numRepeats  \cr
   \tab \cr
   \code{rawDCV} \tab The raw log-ratio scoring matrix with scores from each metric (Hinton(2021)) unaggreagted \cr
}
use_pipe(export = TRUE)
}
\description{
A sparse non parametric association test for compositional/omics data  (Hinton (2021))
}
\examples{
selectionEnergy.scaled()

}
\references{
Hinton, A.L., Mucha, P.J., (2021). Simultaneous variable selection and group association testing in sparse high dimensional omics data. XXXX.
}
