% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{simFromExpData.covarianceShift}
\alias{simFromExpData.covarianceShift}
\title{Simulate large covaraince shift from empirical data}
\usage{
simFromExpData.covarianceShift(
  raMatrix,
  n1 = 30,
  n2 = 30,
  maxCov = 32,
  seed = 8272008,
  perFixedFeatures = 0.8,
  shiftPercent = 1.25
)
}
\arguments{
\item{raMatrix}{empirical relative abundance matrix}

\item{n1}{number of samples class 1}

\item{n2}{number of samples class 2}

\item{maxCov}{xxx}

\item{perFixedFeatures}{xxx}

\item{shiftPercent}{xxx}
}
\value{
A data frame with simulated data
}
\description{
Simulate  covariance shift in data assuming  count data are distributied as an additive logistics normal distribution on the simplex (see Aitichinson (1986))
}
\examples{
simFromExpData.covarianceShift()

}
\references{
Aitchinson, J 1986
}
