% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{simFromExpData.largeMeanShft}
\alias{simFromExpData.largeMeanShft}
\title{Simulate location shift in mean vector of Additive logistics normal distribution}
\usage{
simFromExpData.largeMeanShft(
  raMatrix,
  n1 = 30,
  n2 = 30,
  perFixedFeatures = 0.95,
  featureShiftPercent = 1.25,
  seed = 8272008
)
}
\arguments{
\item{raMatrix}{empirical relative abundance matrix}

\item{n1}{number of samples class 1}

\item{n2}{number of samples class 2}

\item{perFixedFeatures}{xxx}

\item{shiftPercent}{xxx}
}
\value{
A data frame with simulated data
}
\description{
Simulate  location shift in data assuming  count data are distributied as an additive logistics normal distribution on the simplex (see Aitichinson (1986))
}
\examples{
simFromExpData.largeMeanShft()

}
\references{
Aitchinson, J 1986
}
