% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioProcessFunctions.R
\name{stageData}
\alias{stageData}
\title{Wrapper function for processing count/relative data for log-ratio analysis}
\usage{
stageData(featureMatrix, labels, permuteLabel = F, permuteFeatures = F)
}
\arguments{
\item{featureMatrix}{a n-samples x p-features data.frame with count or relative abundance data}

\item{labels}{class labels for data}

\item{permuteLabel}{should lass labels be permuted}

\item{permuteFeatures}{should features be permuted}
}
\value{
A list containing:\tabular{ll}{
   \code{allData} \tab a data.frame where the first column contains class labels and the remaining p columns are features \cr
   \tab \cr
   \code{y_train} \tab A vector of class lables \cr
}
}
\description{
Close data and Impute zeroes. Allow permuting labels or columns for permutation distribution analysis
}
\examples{
\dontrun{
stageData()
}


}
