% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Uguanda_PIH}
\alias{Uguanda_PIH}
\title{16S sequenced cerebrospinal fluid (CSF) case-control dataset}
\format{
A data frame with 92 rows (samples) and 369 variables( Status[pos (PIH) / neg (NPIH)] + OTUs[full taxonomic assignment] ):
}
\source{
\url{https://microbiomedb.org/mbio/app/record/dataset/DS_953b8ff2d4}
}
\usage{
Uguanda_PIH
}
\description{
A microbiome dataset containing the case-control status and Operational Taxonomic Unit (OTU) count composition of infants with/without Postinfectious hydrocephalus (PIH).
Dataset was extracted from the microbiomeDB release 22 and pre-processed in:
 \url{https://github.com/andrew84830813/SelectionEnergyPerm_Project.git}.
}
\keyword{datasets}
