% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{simFromExpData.covarianceShift}
\alias{simFromExpData.covarianceShift}
\title{Simulate large covaraince shift from empirical data}
\usage{
simFromExpData.covarianceShift(
  raMatrix,
  n1 = 30,
  n2 = 30,
  maxCov = 32,
  seed = 8272008,
  perFixedFeatures = 0.8,
  shiftPercent = 1
)
}
\arguments{
\item{raMatrix}{empirical relative abundance matrix}

\item{n1}{number of samples class 1}

\item{n2}{number of samples class 2}

\item{maxCov}{controls the dispersion of \eqn{U} in the formation of the covariance matrix see Hinton(2021)}

\item{seed}{random seed}

\item{perFixedFeatures}{percent of lowest variance features to shift mean vector}

\item{shiftPercent}{percent to shift mean vector of the 'perFixedFeatures' number of features}
}
\value{
A data frame with simulated data
}
\description{
Simulate  covariance shift in data assuming  count data are distributied as an additive logistics normal distribution on the simplex (see Aitichinson (1986))
}
\examples{
\dontrun{
simFromExpData.covarianceShift()
}

}
\references{
Aitchinson, J 1986
}
